# Make TracerModel thread parallel