@bska 2025-04-25
PR #6037 did some early exploration in this area.  Is this PR in any way related to that work?