@akva2 2025-04-25
mostly time. i've been working on the data structures to improve efficiency but consistently gotten side tracked by other things (support tickets, release, ..)