@jakobtorben 2025-04-25
Ahh I see, sorry for unawarly duplicating your work! We should aim to get one of them in then. Getting better thread parallelisation is important for when I do MPI serial runs with GPU acceleration.