@jakobtorben 2025-04-25
We can also merge yours since you added this first. It doesn't matter to me. The most important is to get this in as I need the functionality now. 

But yeah I agree it still not ideal. It is still far away from the MPI parallel pre post part, so there are room for more improvements:

![image](https://github.com/user-attachments/assets/2b00c55b-cdaf-4d1e-a380-034d44f8b4d3)