@jakobtorben 2025-04-25
Ok. Again sorry for duplicating, this was not my intention. I think I need to turn on notifications to stay on top of all new PRs...

I see that you have some other draft branches for thread parallelisation as well. Looking forward to further improvements.

Also if you see some more improvements here, please add on top. As you see from the comparison with MPI parallelisation there is still lots of potential.